Resolve several oversights since alpha9
QAppEventFilter:
 * ignore any event we do not require
 * return a boolean

MainWindow:
 * .show() must be called after the event filter setup, otherwise
   init callbacks are never called nor resolved
 * WatchDog handlers must be set before schedule, wrote a failure state
 * QSettings can't have Mainwindow as parent (I DON'T KNOW WHY)
 * disable _connection_link at init, once the settings are setup at
   first run

+++ qmm/config.py
@@ -17,6 +17,10 @@ logger = logging.getLogger(__name__)
 dirs = appdirs.AppDirs(appname='qmm', appauthor=False)
 
 
+class SettingsNotSetError(Exception):
+    pass
+
+
 def get_config_dir(filename=None, extra_directories=None) -> str:
     """Return the full path of the user config dir.
 
@@ -150,8 +154,10 @@ class Config(MutableMapping):
                 os.fsync(fp)  # noqa
 
             filename = os.path.realpath(filename)
-            shutil.move(filename, "{}.bak".format(filename))
             os.makedirs(os.path.dirname(filename), exist_ok=True)
+            if os.path.exists(filename):
+                shutil.move(filename, "{}.bak".format(filename))
+            logger.debug("Saving new config to %s", filename)
             shutil.move(filename_tmp, filename)
         except IOError as e:
             logger.error("An error occured while saving the settings:\n%s", e)
